requirements.txt
maturin==1.9.6
pytest==8.4.2
pytest-asyncio==1.2.0
psutil==7.1.1
python-dotenv==1.1.1